% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposetable.R
\name{transposetable}
\alias{transposetable}
\title{Transpose a data table: convert rows to columns and columns to rows.
Row names will become column names, and column names will become row names.}
\usage{
transposetable(fnamein, fnameout, wdir = getwd())
}
\arguments{
\item{fnamein}{character vector that specifies the name of
tab-delimited text file with the input data table.}

\item{fnameout}{character vector that specifies the name of output file
where the transposed data table will be saved.}

\item{wdir}{character vector that specifies the name of the working
directory for the input/output files (defaults to the current R directory).}
}
\value{
no return value
}
\description{
Transpose a data table: convert rows to columns and columns to rows.
Row names will become column names, and column names will become row names.
}
\examples{

# Example with data files included in the package:

library(data.table)
library(kmcut)

# Load example gene expression data table for 2 genes
fdat <- system.file("extdata", "example_genes_295.txt", package = "kmcut")

transposetable(fnamein = fdat, fnameout = "example_genes_295_transposed.txt")

# This will create in the current working directory a tab-delimited text
# file with the transposed table: "example_genes_295_transposed.txt"
}
